Handle cross-streets
