+++ lib/ny-addressor/parsers/generic_parser.rb
@@ -3,6 +3,7 @@ require_relative '../address_part'
 require_relative '../../extensions/string'
 require_relative '../constants'
 require_relative '../address_field'
+require_relative '../utils'
 require 'byebug'
 
 module NYAddressor
