+++ lib/ny-addressor/utils.rb
@@ -1,0 +1,31 @@
+module NYAddressor
+  # Fetches constatns for specific region and  type
+  # @param region [Symbol] :US, :UK, etc
+  # @param type [Symbol] :STREET_LABELS, :UNIT_DESCRIPTORS, etc
+  # @return [Hash|List]
+  def self.constants(region, type)
+    begin
+      region_map = Constants.const_get(type)[region]
+      return region_map if region_map
+    rescue
+    end
+    Constants::Generics.const_get(type)
+  end
+
+  # Normalize a string for a given region
+  def self.normalize(input, region)
+    result = input.dup
+    types_to_process = %i[COUNTRY_IDENTIFIERS STATES STREET_NUMBERS STREET_DIRECTIONS STREET_LABELS UNIT_TYPES]
+    types_to_process.each do |type|
+      constants(region, type).each do |full_string, abbreviation|
+        result.gsub!(/\b#{full_string}\b/i, abbreviation)
+      end
+    end
+    remove_cross_street(result)
+  end
+
+  # 1505 & 1510 -> 1505
+  def self.remove_cross_street(input)
+    input.gsub(/(\d+)\s*&\s*\d+/, '\1')
+  end
+end
